Update LotteryPicks_PyQt5
Updates to create the ability to have up to 5 different number combinations, with error validations to ensure the right information is selected before the calculations are done.  This will allow you to run through the random number generator any number of times and then will select the top 5, 4, 3, or last ball that has the most selections to display for the user.  It's kind of fun...